# Speed up the send command