--- a/mycroft/messagebus/send.py
+++ b/mycroft/messagebus/send.py
@@ -15,13 +15,7 @@
 import sys
 import json
 
-from websocket import create_connection
-
-from mycroft.configuration import Configuration
-from mycroft.configuration.locations import (DEFAULT_CONFIG, SYSTEM_CONFIG,
-                                             USER_CONFIG)
-from mycroft.messagebus.client import MessageBusClient
-from mycroft.messagebus.message import Message
+from .send_func import send
 
 
 def main():
@@ -55,36 +49,6 @@ def main():
         exit()
 
     send(message_to_send, data_to_send)
-
-
-def send(message_to_send, data_to_send=None):
-    """Send a single message over the websocket.
-
-    Args:
-        message_to_send (str): Message to send
-        data_to_send (dict): data structure to go along with the
-            message, defaults to empty dict.
-    """
-    data_to_send = data_to_send or {}
-
-    # Calculate the standard Mycroft messagebus websocket address
-    config = Configuration.get([DEFAULT_CONFIG,
-                                SYSTEM_CONFIG,
-                                USER_CONFIG],
-                               cache=False)
-    config = config.get("websocket")
-    url = MessageBusClient.build_url(
-        config.get("host"),
-        config.get("port"),
-        config.get("route"),
-        config.get("ssl")
-    )
-
-    # Send the provided message/data
-    ws = create_connection(url)
-    packet = Message(message_to_send, data_to_send).serialize()
-    ws.send(packet)
-    ws.close()
 
 
 if __name__ == '__main__':