## Description
The `mycroft.messagebus.send` utility is used to send simple messages to the messagebus, but running is unnecessarily slow and affects the commandline tools `mycroft-speak` and `mycroft-say-to`.

This PR removes the fetch of configuration from the backend since the backend should never provide any information on the messagebus setup.

Timing from dev:
```
 $ time python -m mycroft.messagebus.send "speak" "{\"utterance\": \"hello\"}"

real    0m2,700s
user    0m1,000s
sys     0m0,266s
```

Timing from PR branch:
```
 $ time python -m mycroft.messagebus.send "speak" "{\"utterance\": \"hello\"}"

real    0m0,977s
user    0m0,871s
sys     0m0,299s
```

I also added a bit of restructuring removing the annoying warning

```
/usr/lib/python3.6/runpy.py:125: RuntimeWarning: 'mycroft.messagebus.send' found in sys.modules after import of package 'mycroft.messagebus', but prior to execution of 'mycroft.messagebus.send'; this may result in unpredictable behaviour
  warn(RuntimeWarning(msg))
```

I couldn't come up with a better name than `_send.py` but I'll happily accept suggestions.

## How to test
Check out run a send command and verify that things still work as intended.

## Contributor license agreement signed?
CLA [ Yes ]