@davidwagnerkc 2019-10-26
👍

On Sat, Oct 26, 2019 at 1:28 AM Åke <notifications@github.com> wrote:

> or perhaps send_func.py?
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/MycroftAI/mycroft-core/pull/2371?email_source=notifications&email_token=AE4RVZ2RE5QKAGQH4VI3VUDQQPPQXA5CNFSM4JFCTHKKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOECKBFRA#issuecomment-546575044>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AE4RVZ6HXYVJMJAO4SH6IPDQQPPQXANCNFSM4JFCTHKA>
> .
>