@forslund 2019-10-28
A third option is to create a send submodule

mycroft/messagebus/send.py -> mycroft/messagebus/send/__main__.py
mycroft/messagebus/_send.py -> mycroft/messagebus/send/send.py