@forslund 2019-10-29
Ok, going with the send_func.py since all names seem equally terrible